app/requirements.txt
coveralls
flake8
nltk>=3.2,<4.0
nose
Pint==0.18
python-dateutil<2.8,>=2.7
pymongo>=3.3,<4.0
twine
twython
spacy
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.0/en_core_web_sm-2.2.0.tar.gz#egg=en_core_web_sm
sphinx>=3.0,<3.1
sphinx_rtd_theme
PyYAML<4.0,>=3.12
mathparse>=0.1,<0.2
sqlalchemy==1.2.19
pytz==2022.1
docutils<0.18
newspaper3k == 0.2.8